16bit address bus
git-svn-id: svn+ssh://svn.phoenixbox.net/svn/v6502/trunk@17 96dfbd92-d197-e211-9ca9-00110a534b34

+++ v6502/main.c
@@ -53,7 +53,7 @@ int main(int argc, const char * argv[])
 			popArg(command, MAX_COMMAND_LEN);
 			
 			// Make sure we don't go out of bounds either direction
-			uint8_t start = strtol(command, NULL, 16);
+			uint16_t start = strtol(command, NULL, 16);
 			if (start <= 0x10) {
 				start = 0x00;
 			}
